+++ plugins/elevenlabs/src/index.ts
@@ -4,6 +4,7 @@
 import { Plugin } from '@livekit/agents';
 
 export * from './models.js';
+export * from './stt.js';
 export * from './tts.js';
 
 class ElevenLabsPlugin extends Plugin {